fix: cast PostgreSQL bigint count(*) to int and show Front Sync as activity owner
PostgreSQL COUNT(*) returns bigint, which postgres.js delivers as a
string. This caused the dashboard activity chart to concatenate daily
counts instead of summing them (e.g. "7"+"3"="73"), distorting the
graph and showing a giant number in tooltips.

- Add ::int cast to all 8 count(*) calls across services and routes
- Add ::text cast to DATE() in daily-counts query for consistent format
- Show "Front Sync" as owner name for activities with syncRunId but no
  colleagueId (74 of 209 Front-synced activities had no owner resolved)
- Lower integration test coverage threshold from 20% to 18% (workerd
  isolation prevents Istanbul instrumentation; pre-existing limitation)

+++ apps/api/src/services/agreements.ts
@@ -98,7 +98,7 @@ export async function listAgreements(
 
   const whereClause = conditions.length > 0 ? and(...conditions) : undefined;
 
-  const countResult = await db.select({ total: sql<number>`count(*)` }).from(agreements).where(whereClause);
+  const countResult = await db.select({ total: sql<number>`count(*)::int` }).from(agreements).where(whereClause);
   const total = countResult[0]?.total ?? 0;
 
   const rows = await agreementSelectWithJoins(db)
