+++ apps/api/src/services/general-leads.ts
@@ -67,7 +67,7 @@ export async function listGeneralLeads(
 
   const whereClause = conditions.length > 0 ? and(...conditions) : undefined;
 
-  const countResult = await db.select({ total: sql<number>`count(*)` }).from(generalLeads).where(whereClause);
+  const countResult = await db.select({ total: sql<number>`count(*)::int` }).from(generalLeads).where(whereClause);
   const total = countResult[0]?.total ?? 0;
 
   const rows = await db
