+++ apps/api/src/services/lead-scores.ts
@@ -305,7 +305,7 @@ export async function listLeadScores(
 
   const whereClause = conditions.length > 0 ? and(...conditions) : undefined;
 
-  const countResult = await db.select({ total: sql<number>`count(*)` }).from(leadScores).where(whereClause);
+  const countResult = await db.select({ total: sql<number>`count(*)::int` }).from(leadScores).where(whereClause);
   const total = countResult[0]?.total ?? 0;
 
   const rows = await db
